requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
Django==1.11
django-cors-headers==2.0.2
django-filter==1.0.2
django-picklefield==0.3.2
djangorestframework==3.6.2
kombu==4.0.2
Markdown==2.6.8
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
vine==1.1.3
